Testing for Security
As always: we need to start by writing a test. In ProgrammerControllerTest, add a new public function testRequiresAuthentication().

Let's make an API request to an endpoint that should be secured and then assert some things. Start with {dollar_sign}response = {dollar_sign}this->client->post('/api/programmers'). Send this a valid JSON body.

Ok, if our API client tries to anonymously access a secured endpoint, what should be returned? Well, at the very least, assert that the response status code is 401, meaning 'Unauthorized'.

Ok! Let's go make sure this fails! Copy the method name and find the terminal. Run:

./vendor/bin/phpunit --filter testRequiresAuthentication src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
It fails with a validation error: it is getting beyond the security layer and executing our controller. Time to lock that down!