Securing a Controller
Open ProgrammerController. How can we require the API client to be authenticated? The exact same way you do in a web application. Add ->denyAccessUnlessGranted('ROLE_USER').

That's it. I'm using ROLE_USER because all of my users have this role - you could also use IS_AUTHENTICATED_FULLY.

Ok, back to the test! Run it!

./vendor/bin/phpunit --filter testRequiresAuthentication src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
Oh, interesting - it's a 200 status code instead of 401. Look closely: it redirected us to the login page. So, it's kind of working... you can't add programmers anonymously anymore. But clearly, we've got some work to do.