The new Token Resource
Everything we've built so far has been centered around the Programmer resource. Now, we'll be sending back tokens: and you can think of a Token as our second API resource: the client will be able to create new tokens, and potentially, we could allow them to delete tokens.

As always, we'll start with the test. Create a new class called TokenControllerTest. Make it extend the handy ApiTestCase that we've been working on. Add public function testPOSTCreateToken().

Ok, let's think about this. First, we're going to need a user in the database before we start. To create one, add {dollar_sign}this->createUser() with weaverryan and the super-secure and realistic password I<3Pizza.
Next, make the POST request: {dollar_sign}response = {dollar_sign}this->client->post() to /api/tokens.

That URL could be anything, but the most important thing is that it's consistent with the /api/programmers we already have.

The last thing we need to do is send the username and password. And really, you can do this however you want. But, why not take advantage of the classic HTTP Basic Authentication. To send an HTTP Basic username and password with Guzzle, add an auth option and set it to an array containing the username and password.

And hey, reminder time! On production, you will make your API work over HTTPS. The last thing we want is plain-text password flying all over the interwebs.

Below, assert that we get back a 200 status code, or you could use 201 - since technically a resource is being created.

Now, what should the response look like? Well, it should be a token resource... which is really just a string. Use the asserter to assert that the JSON at least contains a token property - we don't know exactly what its value will be.

Looks cool! Copy the method name and run only this test:

./vendor/bin/phpunit src/AppBundle/Tests/Controller/Api/TokenControllerTest.php
This should fail... and it does! A 404 not found. Time to bring this to life!