Create a new TokenController in the Api directory.
Make this extend the same BaseController from our project and let's get to work!

First create a public function newTokenAction(). Add the @Route above and let it autocomplete so that the use statement is added for the annotation. Set the URL to /api/tokens. Heck, let's get crazy and also add @Method: we only want this route to match for POST requests.

To start, don't get too fancy: just return a new Response from HttpFoundation with TOKEN!

Got it! That won't make our test pass, but it is an improvement. Re-run it:

./vendor/bin/phpunit src/AppBundle/Tests/Controller/Api/TokenControllerTest.php
Still failing - but now it has the 200 status code.