Checking the Username and Password
Head back to TokenController. Here's the process:

  1.Check that the username and password are correct.
  2.Generate a JSON web token.
  3.Send it back to the client.
  4.High-five everyone at your office. I can't wait to get to that step.
Type-hint a new argument with Request to get the request object

Next, query for a User object with the normal {dollar_sign}user = {dollar_sign}this->getDoctrine()->getRepository('AppBundle:User') and findOneBy(['username' => '']). Get the HTTP Basic username string with {dollar_sign}request->getUser().

And what if we can't find a user? Throw a {dollar_sign}this->createNotFoundException().
If you wanted to hide the fact that the username was wrong, you can throw a BadCredentialsException instead - you'll see me do that in a second.

Checking the password is easy: {dollar_sign}isValid = {dollar_sign}this->get('security.password_encoder') ->isPasswordValid(). Pass it the {dollar_sign}user object and the raw HTTP Basic password string: {dollar_sign}request->getPassword()

If this is not valid, throw a new BadCredentialsException. We're going to talk a lot more later about properly handling errors so that we can control the exact JSON returned. But for now, this will at least kick the user out.