Ok, ready to finally generate that JSON web token? Create a {dollar_sign}token variable and set it to {dollar_sign}this->get('lexik_jwt_authentication.encoder')->encode() and pass that any array of information you want to store in the token. Let's store ['username' => {dollar_sign}user->getUsername()] so we know who this token belongs to.
 Tip
Don't forget to pass an exp key to the token, otherwise the token will never expire! We forgot to do this in the video!

But you can store anything here, like roles, user information, some poetry - whatever!

And that's it! This is a string, so return a new JsonResponse with a token field set to {dollar_sign}token.

That's it, that's everything. Run the test!
./vendor/bin/phpunit src/AppBundle/Tests/Controller/Api/TokenControllerTest.php

It passes!