We already added a denyAccessUnlessGranted() line to ProgrammerController::newAction(). That means this endpoint is broken: we don't have an API authentication system hooked up yet.
Open up ProgrammerControllerTest() and find testPOST(): the test for this endpoint.

Rename this to testPOSTProgrammerWorks() - this will make its name unique enough that we can run it alone.

Copy that name and run it:

./vendor/bin/phpunit --filter testPOSTProgrammerWorks src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
Instead of the 201, we get a 200 status code after being redirected to /login. I know we don't have our security system hooked up yet, but pretend that it is hooked up and working nicely. How can we update the test to send a token?