getCredentials()
First: getCredentials(). Our job is to read the Authorization header and return the token - if any - that's being passed. To help with this, we can use an object from the JWT bundle we installed earlier: {dollar_sign}extractor = new AuthorizationHeaderTokenExtractor(). Pass it Bearer - the prefix we're expecting before the actual token - and Authorization, the header to look on.

Grab the token with {dollar_sign}token = {dollar_sign}extractor–>extract() and pass it the {dollar_sign}request.

If there is no token, return null.

This will cause authentication to stop. Not fail, just stop trying to authenticate the user via this method.

If there is a token, return it!