getUser()
Next, Symfony will call getUser() and pass this token string as the  argument. Our job here is to use that token to find the user it relates to.

And this is where JSON web tokens really shine. Because if we simply decode the token, it will contain the username. Then, we can just look it up in the database.

To do this, we'll need two services. On top of the class, add a __construct() method so we can inject these. First, we need the lexik encoder service. Go back to your terminal and run:

./bin/console debug:container lexik
Select the lexik_jwt_authentication.encoder service. Ah, this is just an alias for the first service - lexik_jwt_authentication.jwt_encoder. And this is an instance of JWTEncoder. Back in the authenticator, use this as the type-hint. Or wait, since it looks like there's an interface this probably implements, you can use JWTEncoderInterface instead. Give this one more argument: EntityManager {dollar_sign}em.
I'll use a shortcut - option+enter on a Mac - to initialize these fields.

This created the two properties and set them for me. Nice!

Head back down to getUser(). First: decode the token. To do that, {dollar_sign}data = {dollar_sign}this–>jwtEncoder->decode() and pass it {dollar_sign}credentials - that's our token string.

That's it!  is now an array of whatever information we originally put into the token. Fundamentally, this works just like a normal json_decode, except that the library is also checking to make sure that the contents of our token weren't changed. It does this by using our private key. This guarantees that nobody has changed the username to some other username because they're a jerk. Encryption is amazing.

It also checks the token's expiration: our tokens last 1 hour because that's what we setup in config.yml:

 77 lines  app/config/config.yml
... lines 1 - 72
lexik_jwt_authentication:
... lines 74 - 76
    token_ttl:        3600

So, if ({dollar_sign}data === false), then we know that there's a problem with the token. If there is, throw a new CustomUserMessageAuthenticationException() with Invalid token.

In version 2 of the bundle, you should instead use a try-catch around this line:

use Lexik\Bundle\JWTAuthenticationBundle\Exception\JWTDecodeFailureException;
// ...

public function getUser({dollar_sign}credentials, UserProviderInterface {dollar_sign}userProvider)
{
    try {
        {dollar_sign}data = {dollar_sign}this->jwtEncoder->decode({dollar_sign}credentials);
    } catch (JWTDecodeFailureException {dollar_sign}e) {
        // if you want to, use can use ->getReason() to find out which of the 3 possible things went wrong
        // and tweak the message accordingly
        // https://github.com/lexik/LexikJWTAuthenticationBundle/blob/05e15967f4dab94c8a75b275692d928a2fbf6d18/Exception/JWTDecodeFailureException.php

        throw new CustomUserMessageAuthenticationException('Invalid Token');
    }

    // ...
}
We'll talk about what that does in a second.

But if everything is good, get the username with {dollar_sign}username = {dollar_sign}data['username']

Then, query for and return the user with return {dollar_sign}this–>em–>getRepository('AppBundle:User')–>findOneBy() passing username set to {dollar_sign}username