Sending the Authorization Header Everywhere
Let's fix that with as little work as possible. Copy the {dollar_sign}token = code and delete it.

Click into ApiTestCase and add a new protected function called getAuthorizedHeaders() with two arguments: a {dollar_sign}username and an optional array of other {dollar_sign}headers you want to send on the request.

Paste the {dollar_sign}token = code here and add a new Authorization header that's equal to Bearer and then the token. Return the entire array of headers.

Now, copy the method name. Oh, and don't forget to actually use the {dollar_sign}username argument! In ProgrammerControllerTest, add a headers key set to {dollar_sign}this->getAuthorizedHeaders('weaverryan')

And we just need to repeat this on every single method inside of this test. I'll look for {dollar_sign}this->client to find these... and do it as fast as I can!