By hooking into Guzzle, we could add the Authorization header to every request automatically... but there might be some requests where we do not want this header.
In fact, at the bottom, we actually test what happens when we don’t send the Authorization header. Skip adding the header here.

With any luck, we should get a bunch of beautiful passes.

./vendor/bin/phpunit tests/AppBundle/Controller/Api/ProgrammerControllerTest.php
And we do! Ooh, until we hit the last test! When we don't send an Authorization header to an endpoint that requires authentication... it's still returning a 200 status code instead of 401. When we kick out non-authenticated API requests, they are still being redirected to the login page... which is clearly not a cool way for an API to behave.

Time to fix that.