One Entry Point per Firewall
Open up security.yml:
security:
... lines 2 - 8
  firewalls:
    main:
      pattern: ^/
      anonymous: true
      form_login:
        # The route name that the login form submits to
        check_path: security_login_check
        login_path: security_login_form
      logout:
        # The route name the user can go to in order to logout
        path: security_logout
      guard:
        authenticators:
          - 'jwt_token_authenticator'

Here's the problem: we have a single firewall. When an anonymous request accesses the site and hits a page that requires a valid user, Symfony has to figure out what one thing to do. If this were a traditional app, we should redirect the user to /login. If this were an API, we should return a 401 response. But our app is both: we have an HTML frontend and API endpoints. Symfony doesn't really know what one thing to do.
security:
... lines 2 - 8
  firewalls:
    main:
... lines 11 - 12
      form_login:
        # The route name that the login form submits to
        check_path: security_login_check
        login_path: security_login_form
... lines 17 - 3

The form_login authentication mechanism has a built-in entry point and it is taking priority. Our cute start() entry point function is being totally ignored.

But no worries, you can control this! You could add an entry_point key under your firewall and point to the authenticator service to say 'No no no: I want to use my authenticator as the one entry point'. But then, our HTML app would break: we still want users on the frontend to be redirected.

Normally, I'm a big advocate of having a single firewall. But this is a perfect use-case for splitting into two firewalls: we really do have two very different authentication systems at work.