Above, the main firewall, add a new key called api: the name is not important. And set pattern: ^/api/
That's a regular expression, so it'll match anything starting with /api/. Oh, and when Symfony boots, it only matches and uses one firewall. Going to /api/something will use the api firewall. Everything else will match the main firewall. And this is exactly what we want.

Add the anonymous key: we may still want some endpoints to not require authentication.

I'll also add stateless: true. This is kind of cool: it tells Symfony to not store the user in the session. That's perfect: we expect the client to send a valid Authorization header on every request.

Move the guard authenticator up into the api firewall.

And that should do it! Now, it will use the start() method from our authenticator.

Give it a try!

./vendor/bin/phpunit –filter testRequiresAuthentication src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
It passes! Don't rush into having multiple firewalls, but if you have two very different ways of authentication, it could be useful.