ApiProblem in start()
Head to the JwtTokenAuthenticator. In start(), create a new {dollar_sign}apiProblem = new ApiProblem(). Pass it a 401 status code with no type.

The detail key should tell the API client any other information about what went wrong. And check this out: when the start() method is called, it has an optional  argument. Most of the time, when Symfony calls start() its because an AuthenticationException has been thrown. And this class gives us some information about what caused this situation.

And in fact, in TokenController, we're throwing a BadCredentialsException, which is a sub-class of AuthenticationException. Hold command to look inside the class.

It has a getMessageKey() method set to Invalid Credentials.: make sure you test matches this string exactly.

The AuthenticationException - and its sub-classes - are special: each has a getMessageKey() method that you can safely return to the user to help hint as to what went wrong.

Add {dollar_sign}message = {dollar_sign}authException ? {dollar_sign}authException->getMessageKey() : 'Missing Credentials'. If no  is passed, this is the best message we can return to the client. Finish this with {dollar_sign}apiProblem->set('details', {dollar_sign}message)
Finally, return a new JsonResponse with {dollar_sign}apiProblem->toArray() and then a 401.
Perfect! Well, not actually perfect, but it's getting close.

Copy the invalid credentials test method and run:

./vendor/bin/phpunit --filter testPOSTTokenInvalidCredentials src/AppBundle/Tests/Controller/Api/TokenControllerTest.php
It's close! The response looks right, but the Content-Type header is application/json instead of the more descriptive application/problem+json.

Well that's no problem! We just need to set the header inside of the start() method. But wait! Don't do that! Because we've done all of this work before.