Using the new ResponseFactory
We will definitely need this inside ApiExceptionSubscriber, so add it as a second argument: @api.response_factory.

In the class, add the second constructor argument. I'll use option+enter to quickly create that property and set it for me.

Below, it's very simple: {dollar_sign}response = {dollar_sign}this->responseFactory->createResponse() and pass it {dollar_sign}apiProblem.