How Authentication Fails
When JWT authentication fails, what handles that? Well, onAuthenticationFailure() of course.
The getUser() method must return a User object. If it doesn't, then onAuthenticationFailure() is called. In our case, there are two possible reasons: the token might be corrupted or expired or - somehow - the decoded username doesn't exist in our database. In both cases, we are not returning a User object, and this triggers onAuthenticationFailure().

To start, just return a new JsonResponse that says Hello, but with the proper 401 status code.

Copy the testBadToken method name and give it a try!

./vendor/bin/phpunit --filter testBadToken src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
ApiProblem on Failure

It almost works - that's a good start. It proves our code in onAuthenticationFailure() is handling things.