Now, let's setup a proper API problem response, just like we did before: {dollar_sign}apiProblem = new ApiProblem with a 401 status code. Then, use {dollar_sign}apiProblem->set() to add a detail field. And in this case, we always have an AuthenticationException that can hint what went wrong. Use its getMessageKey() method.
Oh, and by the way - if you want, you can send this through the translator service and translate into multiple languages.

Finish this with return {dollar_sign}this–>responseFactory->createResponse() to turn the {dollar_sign}apiProblem into a nice JSON response.

That's it! We did all the hard work earlier.

I want to actually see how this response looks. So, add a {dollar_sign}this->debugResponse() at the end of testBadToken()

Now, re-run the test!

./vendor/bin/phpunit --filter testBadToken src/AppBundle/Tests/Controller/Api/ProgrammerControllerTest.php
Check that out - it's beautiful! It has all the fields it needs, including detail, which is set to Invalid token.

Controlling Error Message

That text is coming from our code, when we throw the CustomUserMessageAuthenticationException. The text - Invalid token - becomes the 'message key' and this exception is passed to onAuthenticationFailure().

This gives you complete control over how your errors look.